scripts/requirements.txt
certifi==2018.1.18
isodate==0.6.0
progressbar2==3.35.2
pyparsing==2.2.0
python-utils==2.3.0
rdflib==4.2.2
six==1.11.0
SPARQLWrapper==1.8.1
